Fix ResetElevatorCommand interruption behavior
- Don't reset encoder when command is interrupted
- Removes RetractElevatorCommand, moving logic into ResetElevatorCommand
